retell-ai-phone-agent/requirements.txt
retell-sdk
python-dotenv